add delete template for user to confirm

+++ products/views.py
@@ -116,7 +116,18 @@ def edit_product(request, product_id):
 
 def delete_product(request, product_id):
     """Delete porduct in the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, f'Product {product.name} has been deleted!')
-    return redirect(reverse('products'))
+    if request.method == 'POST':
+        product.delete()
+        messages.success(request, f'Product {product.name} has been deleted!')
+
+    template = 'products/delete_product.html'
+    context = {
+        'product': product,
+    }
+
+    return render(request, template, context)
